Use index to do search

+++ src/com/localxdata/sql/ExcuteSqlBySingleTable.java
@@ -3,16 +3,10 @@ package com.localxdata.sql;
 
 import java.lang.reflect.Field;
 import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.HashSet;
 
-import com.localxdata.index.IndexTree;
-import com.localxdata.index.IndexUtil;
-import com.localxdata.index.Node;
 import com.localxdata.storage.DataCellList;
 import com.localxdata.storage.StorageNozzle;
 import com.localxdata.struct.DataCell;
-import com.localxdata.util.LogUtil;
 import com.localxdata.util.PraseSqlUtil;
 import com.localxdata.util.XmlUtil;
 import com.localxdata.util.PraseSqlUtil.Action;
@@ -110,13 +104,17 @@ public class ExcuteSqlBySingleTable {
         
         DataCellList deleteList = new DataCellList(); 
         
-        dataList.enterLooper();
-        for(DataCell datacell :dataList) {
-            if(SqlUtil.checkDataByAction(datacell.obj,actionList)) {
-                StorageNozzle.deleteData(tableName, datacell);
-            }   
+        if(SqlUtil.canUseIndex(actionList, tableName)) {
+        	//TODO
+        } else {
+            dataList.enterLooper();
+            for(DataCell datacell :dataList) {
+                if(SqlUtil.checkDataByAction(datacell.obj,actionList)) {
+                    StorageNozzle.deleteData(tableName, datacell);
+                }   
+            }
+            dataList.leaveLooper();
         }
-        dataList.leaveLooper();
     }
     
     public void delete(Object obj) {
